agent-ui: move status out of the reducers folder
