+++ packages/agent-ui/src/containers/processMapsPage.test.js
@@ -7,7 +7,7 @@ import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
 
 import { ProcessMapsPage } from './processMapsPage';
-import * as statusTypes from '../reducers/status';
+import * as statusTypes from '../status';
 
 chai.use(sinonChai);
