+++ packages/agent-ui/src/containers/processSegmentsPage.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux';
 import { withRouter, NavLink } from 'react-router-dom';
 
 import getSegments from '../actions/getSegments';
-import * as statusTypes from '../reducers/status';
+import * as statusTypes from '../status';
 
 export class ProcessSegmentsPage extends Component {
   componentDidMount() {
