+++ packages/agent-ui/src/reducers/maps.test.js
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 
 import maps from './maps';
 import * as actionTypes from '../actions/actionTypes';
-import * as statusTypes from '../reducers/status';
+import * as statusTypes from '../status';
 
 describe('maps reducer', () => {
   it('returns previous state for unknown action', () => {
